pyproject.toml
[project]
name = "ai-secrets-guard"
version = "0.1.0"
description = "Security scanner for AI/LLM projects — secrets detection, prompt injection analysis, vulnerable dependency checks"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "GenaFCS" }]
keywords = ["security", "ai", "llm", "secrets", "prompt-injection", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "anthropic>=0.40",
    "jinja2>=3.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "packaging>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[project.scripts]
ai-secrets-guard = "ai_secrets_guard.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_secrets_guard"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B", "A", "C4", "RET", "SIM"]
ignore = ["S101"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
